+++ stix2validator/test/vulnerability_tests.py
@@ -3,7 +3,7 @@ import copy
 import json
 from . import SCHEMA_DIR
 from .. import validate_string
-from ..validators import ValidationOptions
+from ..util import ValidationOptions
 
 VALID_VULNERABILITY = """
 {
